separated CrudFilter class into its own file; removed FilterCollection class
